requirements-dev.txt
# Development and testing dependencies
# Install with: pip install -r requirements-dev.txt

# Testing frameworks
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.0
pytest-benchmark>=4.0.0
pytest-cov>=4.1.0

# Code quality tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0
pylint>=2.17.0
radon>=6.0.0
xenon>=0.9.0

# Security scanning
safety>=2.3.0
bandit>=1.7.0
pip-audit>=2.6.0
semgrep>=1.45.0

# Documentation
pydocstyle>=6.3.0
interrogate>=1.5.0

# API testing
aioresponses>=0.7.4
responses>=0.23.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Additional useful development tools
ipdb>=0.13.13  # Debugger
pre-commit>=3.3.0  # Git hooks
pytest-xdist>=3.3.0  # Parallel test execution

# Type stubs for better mypy checking
types-requests>=2.31.0
pandas-stubs>=2.0.0